Centralize iframe constants, re-prime closed shadow map after reload
Address ce:review findings:
- Move UNSUPPORTED_IFRAME_SRCS, frame-depth clamping, and the
  ignore-selector resolver into iframe-utils.js so the constants stay
  in sync with other SDKs.
- isUnsupportedIframeSrc lowercases the src so JavaScript:/DATA:
  prefixes are caught.
- Re-call exposeClosedShadowRoots after page.reload() in the responsive
  capture loop — without this the closed-shadow WeakMap is discarded
  with the prior document and subsequent width captures lose closed
  shadow content.
